Migration, Models, Seeder files setup for: Reviews, SpotImages, ReviewImages
